+++ test/clj/lambdacd/presentation/pipeline_state_test.clj
@@ -1,33 +1,52 @@
 (ns lambdacd.presentation.pipeline-state-test
   (:use [lambdacd.testsupport.test-util])
   (:require [clojure.test :refer :all]
-            [lambdacd.presentation.pipeline-state :refer :all]))
+            [lambdacd.presentation.pipeline-state :refer :all]
+            [clj-time.core :as t]))
+
+(def start-time (t/now))
+(def before-start-time (t/minus start-time (t/seconds 10)))
+(def stop-time (t/plus start-time (t/seconds 10)))
+(def after-stop-time (t/plus stop-time (t/seconds 10)))
+
 
 (deftest history-test
   (testing "that it converts the internal pipeline-state into a more readable history-representation"
-    (is (= [{ :build-number 5
-             :status :running}
-            { :build-number 6
-             :status :success}
-            { :build-number 7
-             :status :failure}
-            { :build-number 8
-             :status :waiting}
-            { :build-number 9
-              :status :unknown}
-            { :build-number 10
-             :status :success}
-            ] (history-for { 5  {[0]   { :status :success }
-                                 [1]   { :status :running}}
-                             6  {[0]   { :status :success } }
-                             7  {[0 2] { :status :running}
-                                 [0 1] { :status :failure}}
-                             8  {[0]   { :status :waiting }}
-                             9  {[0]   { :no :status }}
-                             10 {[1 1] { :status :killed }
-                                 [2 1] { :status :success}
-                                 [1]   { :status :success}
-                                 [2]   { :status :success}}})))))
+    (is (= [{:build-number 5
+             :status :running
+             :most-recent-update-at after-stop-time
+             :first-updated-at start-time }
+            {:build-number 6
+             :status :success
+             :most-recent-update-at stop-time
+             :first-updated-at start-time}
+            {:build-number 7
+             :status :failure
+             :most-recent-update-at stop-time
+             :first-updated-at before-start-time}
+            {:build-number 8
+             :status :waiting
+             :most-recent-update-at stop-time
+             :first-updated-at start-time}
+            {:build-number 9
+             :status :unknown
+             :most-recent-update-at stop-time
+             :first-updated-at start-time}
+            {:build-number 10
+             :status :success
+             :most-recent-update-at stop-time
+             :first-updated-at start-time}
+            ] (history-for { 5  {[0]   { :status :success :most-recent-update-at stop-time :first-updated-at start-time }
+                                 [1]   { :status :running :most-recent-update-at after-stop-time :first-updated-at start-time}}
+                             6  {[0]   { :status :success :most-recent-update-at stop-time :first-updated-at start-time}}
+                             7  {[0 2] { :status :running :most-recent-update-at stop-time :first-updated-at start-time}
+                                 [0 1] { :status :failure :most-recent-update-at stop-time :first-updated-at before-start-time}}
+                             8  {[0]   { :status :waiting :most-recent-update-at stop-time :first-updated-at start-time}}
+                             9  {[0]   { :no :status      :most-recent-update-at stop-time :first-updated-at start-time}}
+                             10 {[1 1] { :status :killed  :most-recent-update-at stop-time :first-updated-at start-time}
+                                 [2 1] { :status :success :most-recent-update-at stop-time :first-updated-at start-time}
+                                 [1]   { :status :success :most-recent-update-at stop-time :first-updated-at start-time}
+                                 [2]   { :status :success :most-recent-update-at stop-time :first-updated-at start-time}}})))))
 
 (deftest most-recent-build-test
   (testing "that it returns the most recent build number in the pipeline-state"