+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The official release will have a defined and more stable API. If you are already
 ## 0.2.4
 
 * Bug fixes: #14, #15
-* Minor enhancements: #4
+* Improvements:
+ * Display status in output at the end of a build step (#4)
+ * Display total build time in history (#16)
 
 ## 0.2.3